Short multiplication corrected, Unit tests added

+++ MatrixMultiplier/NumUtils.cs
@@ -10,25 +10,25 @@ namespace MatrixMultiplier
     {
         public static string DecToBin(string dec)
         {
-            int val;
-            if (!int.TryParse(dec, out val)) return String.Empty;
+            long val;
+            if (!long.TryParse(dec, out val)) return String.Empty;
             return Convert.ToString(val, 2);
         }
 
-        internal static int[] DecToBinArray(int A)
+        internal static int[] DecToBinArray(long A)
         {
             string tmp = Convert.ToString(A, 2);
             char[] arr = tmp.Reverse().ToArray();
             return Array.ConvertAll(arr, n => (int)Char.GetNumericValue(n));
         }
 
-        internal static int BinArrayToDec(int[] arr)
+        internal static ulong BinArrayToDec(int[] arr)
         {
-            int res;
+            ulong res;
             res = 0;
             for (int i = arr.Length - 1; i >= 0; i--)
             {
-                res = (res << 1) + arr[i];
+                res = (res << 1) + (ulong)arr[i];
             }
             return res;
         }
